requirements.txt
requests
python-telegram-bot==5.3.0